Move the span related functionality to one place
