+++ compiler/hie-core/hie-core.cabal
@@ -78,15 +78,15 @@ library
     exposed-modules:
         Development.IDE.Logger
         Development.IDE.UtilGHC
-        Development.IDE.Functions.AtPoint
+        Development.IDE.Spans.AtPoint
         Development.IDE.Functions.Compile
         Development.IDE.Functions.CPP
         Development.IDE.Orphans
         Development.IDE.Functions.DependencyInformation
-        Development.IDE.Functions.Documentation
+        Development.IDE.Spans.Documentation
         Development.IDE.Functions.FindImports
         Development.IDE.Functions.GHCError
-        Development.IDE.Functions.SpanInfo
+        Development.IDE.Spans.Calculate
         Development.IDE.Functions.Warnings
         Development.IDE.State.FileStore
         Development.IDE.State.Rules
@@ -102,7 +102,7 @@ library
         Development.IDE.State.Shake
         Development.IDE.Types.Diagnostics
         Development.IDE.Types.Location
-        Development.IDE.Types.SpanInfo
+        Development.IDE.Spans.Type
     other-modules:
         Data.Text.Prettyprint.Doc.Syntax
