+++ compiler/hie-core/src/Development/IDE/Functions/Compile.hs
@@ -25,7 +25,7 @@ import           Development.IDE.Functions.CPP
 import           Development.IDE.Types.Diagnostics
 import qualified Development.IDE.Functions.FindImports as FindImports
 import           Development.IDE.Functions.GHCError
-import           Development.IDE.Functions.SpanInfo
+import           Development.IDE.Spans.Calculate
 import Development.IDE.UtilGHC
 import Development.IDE.Compat
 import Development.IDE.Types.Options
