+++ compiler/hie-core/src/Development/IDE/Spans/Type.hs
@@ -40,14 +40,14 @@ instance Show SpanInfo where
 
 -- we don't always get a name out so sometimes manually annotating source is more appropriate
 data SpanSource = Named Name
-                | Span SrcSpan
+                | SpanS SrcSpan
                 | NoSource
   deriving (Eq)
 
 instance Show SpanSource where
   show = \case
     Named n -> "Named " ++ occNameString (occName n)
-    Span sp -> "Span " ++ show sp
+    SpanS sp -> "Span " ++ show sp
     NoSource -> "NoSource"
 
 getNameM :: SpanSource -> Maybe Name
